requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-environ==0.4.5
django-filter==2.0.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-tables2==2.0.2
djangorestframework==3.8.2
flake8==3.6.0
gunicorn==19.9.0
idna==2.7
marshmallow==2.16.1
mccabe==0.6.1
oauthlib==2.1.0
pycodestyle==2.4.0
pyflakes==2.0.0
python-dotenv==0.9.1
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24
